Update sobolev.lean
there are several issues with these definitions and some of them are wrong. i have added comments letting you know about some problems i see. 

defining holder spaces or working with them will not be used in the parts of Evans we discussed (Ch 2, and Ch 5) until the very end of Chapter 5, in Section 5.6.2, where the Sobolev inequalities for p > n are proved. these issues are very far ahead of where we currently are, and do not enter into any of the current proofs or topics or any parts of the textbook until Section 5.6.2. defining these spaces is not helpful or relevant for the formalization until we intend to formalize the proofs in the p>n case of Sobolev inequalities.  

When we actually get there, these definitions can be written down correctly and easily in ~5 minutes, with the future context that we will then have. I am also confident that by the time we get there, there will have been other changes that will propagate and this file would have to change.